reflected proposal from changseo
