+++ oxide_core/obj_fbx_loader/util.rs
@@ -1,0 +1,9 @@
+use std::fmt::Write;
+
+pub fn join(a: &[&str]) -> String {
+    a.iter()
+    .fold(String::new(),|mut s,&n| {
+        write!(s,"{}",n).ok(); 
+        s
+    })
+}
